globally available cliOptions & configuration
